2007-05-04  Nagappan A  <anagappan@novell.com>
	* libodbc.cs: Wrong ODBC API definitions (using ints instead of
	shorts) - Fixes bug 81306. Thanks to will.mooar@tait.co.nz and
	tausq@debian.org.

	* OdbcDataReader.cs: Wrong ODBC API definitions (using ints
	instead of shorts) - Fixes bug 81306. Thanks to
	will.mooar@tait.co.nz and tausq@debian.org.


svn path=/trunk/mcs/; revision=76657

+++ mcs/class/System.Data/System.Data.Odbc/OdbcDataReader.cs
@@ -904,10 +904,10 @@ namespace System.Data.Odbc
                 {
                         OdbcReturn ret = OdbcReturn.Error;
                         byte [] buffer = new byte [255];
-                        int outsize = 0;
+                        short outsize = 0;
                         int val = 0;
-                        ret = libodbc.SQLColAttribute (hstmt, column, fieldId, 
-                                                       buffer, buffer.Length, 
+                        ret = libodbc.SQLColAttribute (hstmt, (short)column, fieldId, 
+                                                       buffer, (short)buffer.Length, 
                                                        ref outsize, ref val);
                         if (ret != OdbcReturn.Success && ret != OdbcReturn.SuccessWithInfo)
                                 throw new OdbcException (new OdbcError ("SQLColAttribute",
@@ -922,10 +922,10 @@ namespace System.Data.Odbc
                 {
                         OdbcReturn ret = OdbcReturn.Error;
                         byte [] buffer = new byte [255];
-                        int outsize = 0;
+                        short outsize = 0;
                         int val = 0;
-                        ret = libodbc.SQLColAttribute (hstmt, column, fieldId, 
-                                                       buffer, buffer.Length, 
+                        ret = libodbc.SQLColAttribute (hstmt, (short)column, fieldId, 
+                                                       buffer, (short)buffer.Length, 
                                                        ref outsize, ref val);
                         if (ret != OdbcReturn.Success && ret != OdbcReturn.SuccessWithInfo)
                                 throw new OdbcException (new OdbcError ("SQLColAttribute",
